+++ CHANGELOG.md
@@ -1,1 +1,8 @@
 #### 0.0.1.alpha.1
+
+* initial code from community-repo
+
+#### 0.0.1.alpha.2
+
+* add check to get consul checks status of a service
+* add check to get all consul checks in a non-passing state